types: another attempt to resolve conflicts (#396)
* ci(test): run workflow on label assignment (#364)

* ci(test): run workflow on label assignment (#365)

* ci(test): run workflow on label assignment

* ci(test): remove paths

* ci(test): run workflow on label assignment (#367)

* ci(test): run workflow on label assignment

* ci(test): remove paths

* ci(test): run workflow on label assignment

* ci: add "check" workflow (#370)

* ci: add "check" workflow

* ci: add check workflow

* idk

* ci(check): rename to "Type-check"

* refactor(tests): separate tests into categories (#369)

* refactor(tests): seperate tests into categories

* idk

* fix default avatars

* test use depth 10 for error

* remove extra log

* test fix

* test: attempt to fix "Token not provided" error (#384)

* test: fix "Token not provided" error

* idk wtf

* Revert "idk wtf"

This reverts commit 26063aa12746509ba3a1125519991d8b353ad221.

* https://open.spotify.com/track/3ZWc3Bm8eUMRRJhSWKHHeV 🔥🔥

* Revert "test: attempt to fix "Token not provided" error (#384)"

This reverts commit 6f21de2a3a6ed5fcbfbc93541d1faef56b5811d4.

* test(message): add missing tests (#371)

* test(message): add missing tests

* Add todo comment

* docs: add jsdoc comments for event handlers (#376)

* ci: another attempt to fix CI

* ci: run workflow on push to master (#391)

* types(message): make attachment dimensions optional

Reference: https://github.com/discord/discord-api-docs/pull/2214

* types(message): add 16 & 17 to MessageTypes

* these aren't needed

* create: OverwritePayload to Overwrite converter

* fix: type issues
