+++ .github/workflows/test.yml
@@ -1,7 +1,5 @@
 name: Test
-on:
-  pull_request:
-    types: labeled
+on: [push, pull_request]
 jobs:
   test:
     if: github.event.label.name == 'run-tests'
